requirements.txt
google-adk
requests
beautifulsoup4
python-dotenv
google-cloud-firestore
serpapi